Added hard-wired demo of Jobs schema
The command schema:add now does a Propel build hardwired to this schema.
Supports entities so that the user does not have to enter them themselves:
current_header (pk and meta for current record), versionable_header
(pk & meta for versioned record), and known_nodes (local table of
known nodes to fk to).