Some initial code ideas for message signing
Ideally, all messages will be signed, since a static IP is not a
great guarantee that the message is from who we think it is.
Users on a shared server all have a single IP, and we don't want
to have to reauthenticate if a user moves servers. Hence message
signing - openssl preferred, and maybe GPG on the CLI as a second
fall-back, and then IP checking as a third fall-back.

This system will give us encryption as well for very little extra
effort, although I suspect that will be off by default, as it may
add a bit of CPU overhead.