Removed the reverse column for local trusts only
The reverse column in the trust tables is intended to reverse the direction
of the trust. So if we have nodes A and B, with direction=forward it means
that A trusts B. With direction=reverse, it means B trusts A.

However this is only of use for trusts involving a remote node, since A
and B are of different types (and we don't want to have another remote
table!). But for local trusts, if we want to express a reverse trust,
the nodes can be swapped, since A and B are of the same type.