Modify approach to node model foreign key references
Previously where a node schema was processed by the fixup process,
any columns marked as primary keys were made into ordinary
columns, and then a new key was added (row_ord, creator_node_id).
This would have prevented composite primary keys being used in a
schema design.

In that design, any FKs in the schema would have had to have been
entered manually, which is awkward for the user. So, I have
modified the approach so that the user's primary keys are allowed
(thus permitting composite keys) but in each case, a
creator_node_id is also made part of the key. Foreign keys are
automatically modified to contain this extra column, thus
simplifying their specification.

I'm not yet sure this approach is the correct one for Meshing,
but it is a good start. We'll find out soon enough, via unit tests
and setting up experimental meshes.