Ensure a custom base object works with the model unit tests
In the first version we'll just use a custom base Propel class and
capture the pre- and post-save hooks to write the version rows. This
requires an extra select before updates however, so is a bit slower
than capturing the state of an object when it is instantiated and
using that for the version row.

To hook into the creation of a row object will require the writing
of custom code into generated child classes, which will take a bit
more work - so that can be tackled later.