Change versioning approach
Previously, the strategy to save versions was to keep the current
line in the current table, and previous lines in the versionable
table. This creates inconsistencies between an insert and an update,
so I've changed it so every insert/update creates a versioned row.

Where a versioned row is created, the *previous* values are inserted
into the data columns. This means that for an insert, these are all
null - the fixup process has been modified to ensure all versionable
columns permit this.

This happily means that the hash has been dropped from the current
table, since the hash for a particular current row can be read from
the latest versionable row. This is the same for all metadata: the
latest versionable row, whilst the data columns are previous values,
all the metadata columns are current.