Added method to inject version metadata into save
Metadata is now set externally in business logic, rather than getting
this info (e.g. timestamps) from inside the model. This is a better
system, since some info (e.g. whether a change was remote and so has
a received timestamp) is not available in the model.

(It may be available from the model if we implement a changes table,
but for the time being this change feels better.)