Add skeleton for testing the generation of model version numbers
There is a race condition possibility due to the way in which we
generate a MAX() of the current version and then later do a save
operation. I've made a number of changes to:

* encourage a race condition in the test environment
* introduce a table locking system (still to be implemented)
* permit the specification of custom object and peer parent
  classes for models

Some tidying to be done before locking is testable - need to
put the child task log somewhere sensible, zero it before use,
and to apply test methods to its contents. The parent task
needs to wait for all test children to report before exiting
too.