Class changes to better support several hashing strategies
I'd tend towards the view that it would be good to have just one
common hashing strategy for all Meshing instances. However, this
facility should be useful for deciding which is the best one during
the prototyping process. That said, there may be benefit in being
able to customise this, and it may be better to have meshes
operating precisely as its proposers envisage than to force
one-size-fits-all strategies.

If this were to be part of the final version of Meshing, it would
be specified in the schema, and stripped out before Propel gets to
process it. Accordingly the user would not normally need to set
it manually.