Swap local node references for FQDN value when hashing
This is important since two nodes are likely to have a different primary
key for the same known node. We therefore lookup the FQDN of that node
which is guaranteed to be the same between the two.

(This does not yet take account of FQDN changes, which we may have to
deal with later).