Changes in preparation for the Version hash provider tests
In particular, getting the hash for an unsaved row, or for a saved
row where the version row has not yet been written, used to throw
an exception. This has changed to return null, as we can't very
well have exceptions happen in the normal course of things inside
a save (even if they are internally handled).