Set up system to clear and rebuild DatabaseMaps where appropriate
There's some new skeleton code in 4_version_test.php that will test the
protected method to create version numbers. To get this to work, I
regenerated the models with a different class prefix (TestVersion
rather than TestModel) but since the new classes refer to the same
tables as the old ones, having both in the DatabaseMap caused errors
with Propel internally.

To fix this, we needed a method to clear out the existing TableMaps,
which in turn lets the new ones work correctly.

Since in 5_hashing_test.php I go back to the TestModel* prefix, we
need to rebuild the DatabaseMap a second time. This is because this
normally only happens when a model peer is loaded, and these classes
have already been brought into memory as part of previous tests.