Remove PK generator column from system schema table
The original design assumed that there would not be more than one
node having a single schema. That has changed, so if we want a
sequence generator then we'll have to do it per table instance,
rather than per schema table.

For the time being I will rely on PostgreSQL sequences, but we'll
have to see how MySQL fares with this to see if it is necessary.
If it is implemented, perhaps this should go in the node db rather
than the system db, to make node data more easily portable?