Create a 'change table' for each table in the schema
The original design specified a single change table for all nodes and
tables within the nodeset, to be stored in the system database. This
approach was predicated on having a primary key (creator_node_id, ord)
for every table. However this has changed - tables now have
(creator_node_id) added to their existing primary key, which may be
composite.

There were two approaches around this - have a key conversion table
that translates a current table primary key into an integer key
for the change table, or to have one change table per current table.
I've chosen the latter for the time being. I did think that having
one change table per current table would make for more selects, and
so make the system slower, but for large row counts, this may be a
win as tables will be faster to access.

Nodes are more portable between database servers as a result of now
choosing to store the change table with the data.

So - this commit adds change tables, currently unused.