Change build system back to continue on SQL errors
In 9cba4c4112 I'd removed the continue feature when an SQL error is
encountered, but I've just realised why this is required. All table
creation scripts delete tables before they are created, but if they
don't exist this deletion causes an error. I've put this back until
I think of a better way to deal with this.